dns/requirements.txt
dnslib==0.9.7
mysql-connector-python
